+++ src/Base64StringElementTrait.php
