+++ src/ElementInterface.php
@@ -14,6 +14,22 @@ use DOMElement;
 interface ElementInterface
 {
     /**
+     * Get the XML qualified name (prefix:name) of the element represented by this class.
+     *
+     * @return string
+     */
+    public function getQualifiedName(): string;
+
+
+    /**
+     * Test if an object, at the state it's in, would produce an empty XML-element
+     *
+     * @return bool
+     */
+    public function isEmptyElement(): bool;
+
+
+    /**
      * Get the value of an attribute from a given element.
      *
      * @param \DOMElement $xml The element where we should search for the attribute.
