+++ src/Exception/MissingElementException.php
