+++ tests/PHPUnit/ArrayizableElementTestTrait.php
