+++ tests/Utils/StringElement.php
@@ -5,6 +5,7 @@ declare(strict_types=1);
 namespace SimpleSAML\Test\XML;
 
 use SimpleSAML\XML\AbstractElement;
+use SimpleSAML\XML\SerializableElementTrait;
 use SimpleSAML\XML\StringElementTrait;
 
 /**
