+++ tests/resources/xml/domdocument_doctype.xml
