+++ tests/resources/xml/domdocument_invalid_xml.xml
