+++ tests/resources/xml/domdocument_valid_xml.xml
