+++ tests/resources/xml/ssp_ExtendableElement.xml
