+++ tests/src/DOMDocumentFactoryTest.php
