+++ tests/src/StringElementTraitTest.php
